# FamilyClub.React/famiy.club.react/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FamilyClub.WebAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Translators": {
      "get": {
        "tags": [
          "Translators"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Translators"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Translators/{id}": {
      "get": {
        "tags": [
          "Translators"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Translators"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyClub.BLL.DTOs.Translator.TranslatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Translators"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FamilyClub.BLL.DTOs.Translator.TranslatorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "translatorName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}